add `Py_HashBuffer` to `pyo3-ffi` (#5086)
* add `Py_HashBuffer` to `pyo3-ffi`

* Add `Py_HashBuffer` binding when 3.14+

* Move to `cpython/pyhash.rs`

* Update changelog

* Enable`Py_HashBuffer` compat on 3.14 abi3
