+++ newsfragments/5086.added.md
@@ -1,0 +1,3 @@
+- add `Py_HashBuffer` to `pyo3-ffi`
+- add `Py_HashPointer` to `pyo3-ffi`
+- add `PyObject_GenericHash` to `pyo3-ffi`
