+++ pyo3-ffi/src/compat/mod.rs
@@ -52,8 +52,10 @@ macro_rules! compat_function {
 
 mod py_3_10;
 mod py_3_13;
+mod py_3_14;
 mod py_3_9;
 
 pub use self::py_3_10::*;
 pub use self::py_3_13::*;
+pub use self::py_3_14::*;
 pub use self::py_3_9::*;
